requirements.txt
openai==1.3.0
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
statsmodels==0.14.0 